Remove parquet dataset feature flags
Why these changes are being introduced:

Feature flags were used during development work to support
the new parquet dataset ETL architecture and the previous
discrete files architecture.  These flags are no longer needed.

How this addresses that need:
* Removes all code branches and files related to parquet
development feature flagging

Side effects of this change:
* The "v1" behavior that predated the parquet dataset
is no longer supported.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-490
