+++ tests/conftest.py
@@ -13,8 +13,6 @@ def _test_env(monkeypatch):
     monkeypatch.setenv("TIMDEX_ALMA_EXPORT_BUCKET_ID", "test-alma-bucket")
     monkeypatch.setenv("TIMDEX_S3_EXTRACT_BUCKET_ID", "test-timdex-bucket")
     monkeypatch.setenv("WORKSPACE", "test")
-    # NOTE: FEATURE FLAG: remove after v2 work is complete
-    monkeypatch.setenv("ETL_VERSION", "1")
 
 
 @pytest.fixture(autouse=True)
@@ -65,11 +63,3 @@ def s3_client():
 @pytest.fixture
 def run_id():
     return "run-abc-123"
-
-
-# NOTE: FEATURE FLAG: remove after v2 work is complete
-@pytest.fixture
-def etl_version_2(monkeypatch):
-    etl_version = 2
-    monkeypatch.setenv("ETL_VERSION", f"{etl_version}")
-    return etl_version
